Updated Appointmens history component
